补充完善接口 修复仓库 count 方法返回 Long 类型转换的问题

+++ server/src/main/java/org/tcpx/sharine/repository/BookmarkRepository.java
@@ -29,10 +29,10 @@ public interface BookmarkRepository extends JpaRepository<Bookmark, Long> {
     void deleteByUserIdAndVideoId(Long userId, Long videoId);
 
     @Cacheable(unless = "#result==null")
-    Long countByUserId(Long userId);
+    Integer countByUserId(Long userId);
 
     @Cacheable(unless = "#result==null")
-    Long countByVideoId(Long videoId);
+    Integer countByVideoId(Long videoId);
 
 
     List<Bookmark> findByUserId(Long userId, Pageable pageable);
